requirements.txt
discord.py==2.1.1
python-dotenv==0.21.1
revChatGPT==2.3.5
